backend/requirements.txt
# --- Core ---
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic[email]
pydantic-settings>=2.4.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
starlette>=0.38.0

# --- Firebase / Google ---
firebase-admin>=6.5.0
google-api-python-client>=2.140.0
google-auth-oauthlib>=1.2.0
google-cloud-firestore>=2.17.0
google-cloud-storage>=2.18.0
google-generativeai>=0.7.0
google-adk

# --- AI / Media ---
Pillow>=10.0.0

# --- Auth ---
bcrypt>=4.1.0
passlib>=1.7.4
python-jose>=3.3.0
PyJWT>=2.8.0
cryptography>=42.0.0

# --- Database ---
SQLAlchemy>=2.0.30
sqlmodel>=0.0.21

# --- Social Media ---
tweepy>=4.14.0
requests>=2.32.0
requests-oauthlib>=2.0.0
aiohttp>=3.9.0
httpx>=0.27.0

# --- Scheduler ---
APScheduler>=3.10.0

# --- Other ---
jsonschema>=4.20.0
itsdangerous>=2.2.0
